// types/fixed-asset.ts
type Asset = {
  entity_cd: string;
  entity_name: string;
  reg_id: string;
  fa_descs: string;
  div_descs: string;
  dept_descs: string;
  location: string;
  acquire_date: string;
  purchase_date: string;
  descs: string;
  qr_url_attachment: string;
  isprint: string;
  asset_qty: number;
  audit_status: string | null;
  batch_no: number;
  cost_value: number;
  depr_aperiod: number;
  depr_fyear: number;
  depr_method: string;
  depr_rate: number;
  depr_start_aperiod: number;
  depr_start_fyear: number;
  depr_todate: number;
  dept_cd: string;
  div_cd: string;
  doc_date: string;
  doc_no: string;
  fa_cd: string;
  fa_staff_id: string;
  file_name: string | null;
  isqr: string;
  item_cd: string | null;
  level_no: string | null;
  location_cd: string;
  location_map: string | null;
  name: string | null;
  note: string | null;
  old_reg_id: string | null;
  opening_value: number;
  parent_cd: string;
  picture_col: string;
  qualify_cost: number;
  ref_no: string;
  remarks: string;
  residu_value: number;
  room_descs: string;
  serial_no: string;
  source_file_attachment: string | null;
  staff_id: string;
  staff_name: string;
  status: string;
  status_review: string | null;
  supplier_cd: string | null;
  tenure: string;
  trx_type: string;
  unit_price: number;
  url_file_attachment: string | null;
  url_file_attachment2: string | null;
  url_file_attachment3: string | null;
  warranty: number;
  warranty_uom: string;
};

type AssetHistory = {
  entity_cd: string;
  reg_id: string;
  trx_date: string;
  old_location_map: string;
  new_location_map: string;
  old_status_review: string;
  new_status_review: string;
  note: string;
  audit_status: string;
  audit_user: string;
  audit_date: string;
  rowID: number;
};

export type { Asset, AssetHistory };